+++ Controllers/NewsController.cs
@@ -1,4 +1,5 @@
-﻿using BachelorThesis.Models;
+﻿using System.Security.Claims;
+using BachelorThesis.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using BachelorThesis.Database;
@@ -19,15 +20,18 @@ public static class NewsController
             return await db.News.OrderByDescending(s => s.DateAdded).ToListAsync();
         }).RequireAuthorization();
 
-        app.MapPost("api/news/add", async (StudyDb db, News news) =>
+        app.MapPost("api/news/add", async (StudyDb db, News news, HttpContext context) =>
         {
+            var user = db.Users.SingleOrDefault(u => u.Email == context.User.FindFirstValue("preferred_username"));
+            if (user is null) return Results.NotFound();
+
             news.DateAdded = DateTime.Now;
-            news.UserID = 0;
+            news.UserID = user.UserID;
             await db.News.AddAsync(news);
             await db.SaveChangesAsync();
 
             var log = new Log();
-            log.UserID = 0;
+            log.UserID = user.UserID;
             log.Event = LogEvent.NewsAdded;
             log.DateAdded = DateTime.Now;
             log.NewsID = news.NewsID;
@@ -45,14 +49,17 @@ public static class NewsController
             return Results.Ok(news);
         }).RequireAuthorization();
 
-        app.MapPut("api/news/{newsID}/edit", async (StudyDb db, News updatedNews, int newsID) =>
+        app.MapPut("api/news/{newsID}/edit", async (StudyDb db, News updatedNews, int newsID, HttpContext context) =>
         {
+            var user = db.Users.SingleOrDefault(u => u.Email == context.User.FindFirstValue("preferred_username"));
+            if (user is null) return Results.NotFound();
+
             var news = await db.News.FindAsync(newsID);
             if (news is null) return Results.NotFound();
             news.Content = updatedNews.Content;
 
             var log = new Log();
-            log.UserID = 0;
+            log.UserID = user.UserID;
             log.Event = LogEvent.NewsEdited;
             log.DateAdded = DateTime.Now;
             log.NewsID = news.NewsID;
@@ -63,14 +70,17 @@ public static class NewsController
             return Results.Ok();
         }).RequireAuthorization();
 
-        app.MapDelete("api/news/{newsID}/delete", async (StudyDb db, int newsID) =>
+        app.MapDelete("api/news/{newsID}/delete", async (StudyDb db, int newsID, HttpContext context) =>
         {
+            var user = db.Users.SingleOrDefault(u => u.Email == context.User.FindFirstValue("preferred_username"));
+            if (user is null) return Results.NotFound();
+
             var news = await db.News.FindAsync(newsID);
             if (news is null) return Results.NotFound();
             db.News.Remove(news);
 
             var log = new Log();
-            log.UserID = 0;
+            log.UserID = user.UserID;
             log.Event = LogEvent.NewsDeleted;
             log.DateAdded = DateTime.Now;
             log.NewsID = news.NewsID;
